orangepi_rag_service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
numpy>=1.26.0
PyYAML>=6.0.1
sentence-transformers>=2.6.0
